JPERF-902: Print stacktraces on gradle task failure by default
We currently have a problem that the release process on Github Actions is reporting some failure, however the cause it not fully clear.
Showing stacktrace of the task could reveal more information.

Possibly the setting is worth to keep, as we probably always are interested in sourcecode cause when an exception is thrown.
Exceptions seem to usually rely on their stacktraces to be diagnosable. In our case message "https://github.com/atlassian/jira-software-actions: not authorized" doesn't give much context.