JPERF-902: Apply exact instructions for release task from the plugin we are using
The release process is still not working on CI. I still believe it's due to credentials. Apparently the cyrrent way doesn't work with the automated token.

The changed way of passing credentials is based on https://axion-release-plugin.readthedocs.io/en/latest/configuration/ci_servers/#github-actions
This plugin is used as part of gradle-release