JPERF-573: Add visited projects to JQL memory
This change is already present in `JiraCoreScenario`.
It addresses problem when other sources of JQLs are not valid in dataset of tested Jira.
e.g. if we don't have resolved issues.
It transitively also addresses problem when non-JQL sources of issues are not valid.
e.g. if browse boards doesn't return any board or all remembered boards are empty